+++ src/django_prodserver/backends/workers/base.py
@@ -1,0 +1,3 @@
+from ..base import BaseWorkerBackend
+
+__all__ = ["BaseWorkerBackend"]