+++ src/django_prodserver/management/commands/worker.py
@@ -1,0 +1,10 @@
+from ...backends.base import BaseWorkerBackend
+from .server import BaseProcessCommand
+
+
+class Command(BaseProcessCommand):
+    """Start a configured background task worker process."""
+
+    help = "Start a configured background task worker process."
+    process_label = "worker"
+    backend_base_class = BaseWorkerBackend